backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
playwright==1.51.0
requests==2.32.3
opencv-python-headless==4.13.0.92
numpy==2.4.2
gunicorn==24.0.0
google-generativeai==0.8.6
Pillow==11.1.0
earthengine-api==1.5.0
